Revert "Dashboard middle wrapper"
This reverts commit 191f37b0bcfec6899f90ad235f8b91d442902550.
