+++ package.json
@@ -19,7 +19,6 @@
     "font-awesome": "^4.7.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-multi-carousel": "^2.8.0",
     "react-redux": "^8.0.1",
     "react-router-dom": "^6.2.1",
     "react-scripts": "4.0.3",
