+++ src/Assests/Styles/dashboard.page.css
@@ -17,12 +17,3 @@
   font-weight: 500;
   color: #482673;
 }
-
-.mw-heading {
-  color: #023B59;
-  text-transform: uppercase;
-}
-
-.mw-sub-heading {
-  color: #ED0870;
-}
