+++ src/index.js
@@ -5,7 +5,6 @@ import App from './App';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import 'font-awesome/css/font-awesome.css';
 import 'react-toastify/dist/ReactToastify.css';
-import "react-multi-carousel/lib/styles.css";
 import { ToastContainer} from 'react-toastify';
 ReactDOM.render(
   <React.StrictMode>
